+++ packages/ranger-eslint-config/index.js
@@ -425,6 +425,8 @@ module.exports = {
     'react/no-danger': 0,
     'react/function-component-definition': 0,
     'react/require-default-props': 0,
+    'react/state-in-constructor': 0,
+    'react/destructuring-assignment': 0,
     // prettier代码格式化
     'prettier/prettier': [
       'error',