+++ packages/ranger-ui/src/JsonView/index.ts
@@ -1,2 +1,0 @@
-export type { JsonViewProps } from './jsonView'
-export { default as JsonView } from './jsonView'